% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/server.R, R/ui.R
\name{qRAT - qPCR Relative Expression Analysis Tool}
\alias{qRAT - qPCR Relative Expression Analysis Tool}
\alias{qRAT}
\title{Run the shiny application}
\usage{
qRAT(...)
}
\description{
Run the shiny application

Main functions for the qRAT Shiny app

UI for the qRAT Shiny app
}
\author{
Daniel Flatschacher
}
