% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR.R
\name{read.qPCRtable}
\alias{read.qPCRtable}
\title{read qPCR data with HTqPCR and calculate with ddCt packages}
\usage{
read.qPCRtable(fname, na.value = 40, ...)
}
\description{
read qPCR data with HTqPCR and calculate with ddCt packages
}
