% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLink.R
\name{shinyLink}
\alias{shinyLink}
\title{Shiny Link}
\usage{
shinyLink(to, label)
}
\arguments{
\item{to}{page to navigate to}

\item{label}{text that describes the link}
}
\description{
Create a link to an internal tab panel
}
